+++ src/CssDependencyTemplate.js
@@ -1,0 +1,4 @@
+export default class CssDependencyTemplate {
+  // eslint-disable-next-line class-methods-use-this
+  apply() {}
+}
