+++ src/utils.js
@@ -1,0 +1,23 @@
+const MODULE_TYPE = 'css/mini-extract';
+
+function compareIds(a, b) {
+  if (typeof a !== typeof b) {
+    return typeof a < typeof b ? -1 : 1;
+  }
+
+  if (a < b) {
+    return -1;
+  }
+
+  if (a > b) {
+    return 1;
+  }
+
+  return 0;
+}
+
+function compareModulesByIdentifier(a, b) {
+  return compareIds(a.identifier(), b.identifier());
+}
+
+export { MODULE_TYPE, compareModulesByIdentifier };
